requirements.txt
ovh
python-decouple
requests